Moved seven-related things to 'single-file action'
Removed seven logic (and also reconnect logic) into their own folder /
files.  They are completely removed from the monster of a GameController
and other things can be moved out as necessary.
