+++ api/controllers/game/reconnect.js
@@ -1,0 +1,5 @@
+module.exports = function (req, res){
+  Game.subscribe(req, [req.session.game]);
+  sails.sockets.join(req, 'GameList');
+  return res.ok();
+};
