+++ api/controllers/game/seven/jack.js
@@ -1,4 +1,4 @@
-module.exports = function sevenJack(req, res) {
+module.exports = function (req, res) {
   const Promise = require("bluebird");
   const promiseGame = gameService.findGame({gameId: req.session.game});
   const promisePlayer = userService.findUser({userId: req.session.usr});
