+++ api/controllers/game/seven/scuttle.js
@@ -1,5 +1,5 @@
 
-module.exports = function sevenScuttle(req, res) {
+module.exports = function (req, res) {
   const Promise = require("bluebird");
   const promiseGame = gameService.findGame({gameId: req.session.game});
   const promisePlayer = userService.findUser({userId: req.session.usr});
