+++ api/controllers/game/seven/untargeted-one-off.js
@@ -1,5 +1,4 @@
-module.exports = function sevenUntargetedOneOff(req, res) {
-  const Promise = require("bluebird");
+module.exports = function (req, res) {
   const promiseGame = gameService.findGame({gameId: req.session.game});
   const promisePlayer = userService.findUser({userId: req.session.usr});
   const promiseCard = cardService.findCard({cardId: req.body.cardId});
