+++ config/blueprints.js
@@ -44,7 +44,7 @@ module.exports.blueprints = {
   *                                                                          *
   ***************************************************************************/
 
-  // actions: true,
+  actions: false,
 
   /***************************************************************************
   *                                                                          *
