+++ config/config.js
@@ -1,7 +1,4 @@
 
-
-
-
 module.exports = {
 
    orm: {
